+++ Infra.MongoDB/Commons/Helpers/FilterHelper.cs
@@ -40,13 +40,13 @@ public static class FilterHelper
         filterDefinitions.Add(Builders<T>.Filter.Lte(expression, dateTime));
     }
 
-    public static void AddGreaterOrEqualFilter<T>(List<FilterDefinition<T>> filterDefinitions, Expression<Func<T, object>> expression, double value)
+    public static void AddGreaterOrEqualFilter<T>(List<FilterDefinition<T>> filterDefinitions, Expression<Func<T, object>> expression, float value)
     {
         if (value == 0) return;
         filterDefinitions.Add(Builders<T>.Filter.Gte(expression, value));
     }
 
-    public static void AddLessOrEqualFilter<T>(List<FilterDefinition<T>> filterDefinitions, Expression<Func<T, object>> expression, double value)
+    public static void AddLessOrEqualFilter<T>(List<FilterDefinition<T>> filterDefinitions, Expression<Func<T, object>> expression, float value)
     {
         if (value == 0) return;
         filterDefinitions.Add(Builders<T>.Filter.Lte(expression, value));
